Let xmss_core decide on secret key size
This allows different backends to store additional state information
in the secret key while the rest of the codebase remains agnostic.

In particular, this prepares for a common xmss_core.h API for both
the standard and the BDS-traversal-based implementations.

+++ params.h
@@ -26,7 +26,7 @@ typedef struct {
     unsigned int index_bytes;
     unsigned int sig_bytes;
     unsigned int pk_bytes;
-    unsigned int sk_bytes;
+    unsigned long long sk_bytes;
     unsigned int bds_k;
 } xmss_params;
